Make deamon restart-able by listing VMs at init time and major cleanup
